resolver: Use fallback nameservers when none are otherwise supplied
Fallback nameservers are added to the nameserver list with their
interface index set to -1. When a nameserver entry is removed, check
if the fallback ones should be used instead. When nameserver entries
have been added and flushed and at least one of them has an
interface index, remove all fallback nameservers.

Fixes CM-652