gsupplicant: Fix invalid memory access when network interface is removed
There is a possibility that a DBus call to wpa_supplicant is pending
when we shutdown the interface. The interface is shutdown when for
example we disable the wifi technology. When the DBus call returns
results, the interface is already gone and we end up accessing
already freed memory which can cause crashes.

The fix is to cancel all the pending DBus calls when the interface
is removed.