service: Use only nameservers whose IP type is enabled
Use only IPv4 or IPv6 nameservers when the respective IPv4 or IPv6
configuration is enabled. When updating nameservers, first remove
all nameservers in order to add only those that are enabled.

Add only the enabled nameservers to the 'Nameservers' service property
but add all configured nameservers to 'Nameservers.Configuration'
property regardless of the state of the respective IP configuration.