gweb: Handle proxies as addresses and hostnames
It seems the proxy handling code was initially written to only handle
proxies in the form of IPv4 addresses. 38b75abddb5b changed that
implicitly by always doing a hostname lookup for the proxy address,
which fixes proxies given by hostname but breaks IP based proxy
configuration (as sending an A query to most DNS server for an IP
address gets you a result with no answers).

Fix this issue by short-circuiting the resolving step in case the proxy
address is in the form of either an IPv4 or an IPv6 IP literal.