client: Remove extra string duplication when completing commands
Update the command argument completion callback function
__connmanctl_get_lookup_func() to accept the full command line typed
so far. In addition to comparing the input line against each command
in turn, check also that the input line ends with a space after the
command. This allows removal of the string duplication when looking
for the command.

Also, allow completion if there is a command argument completion
callback function. If none found, there is nothing to complete.