service: Remove duplicate EAP AgentPassphrase information
EAP AgentPassphrase has been set on replies from the agent and
removed on any error. It has been used by the wifi plugin as a
backup if the normal wifi passphrase is unset. The normal passphrase
is removed only on invalid key errors with a comment indicating
that EAP does not always signal invalid key errors although the
key is incorrect.

Thus the implementation can end up in a situation where the normal
passphrase is wrong but not removed resulting in unnecessarily
retrying with an erroneous key.