gsupplicant: rely on wpa_supplicant retry mechanism
When connecting WiFi with a wrong passphrase,
connman is longer than it should to cancel the attempt.

The effect is this change, is to have a quicker feedback to the agent
in charge of passphrase input.

Since, Wpa_supplicant 2.0 and later
is already managing its own retry mechanism,
let's use it because it's overkill to have this also inside comman
and even more if it add an unnecessary delay.

Note, that for earlier versions of wpa_supplicant this change may be reverted,
while for later ones (2.0 to HEAD/2013-12-10),
it's resulting a 10 secs improvement
(observed the whole test is now 10s vs 20s between two user's attempts).

Let me credit Liu Guoqiang and Tomasz Bursztyka
for helping to solve this problem.