session: Mark session active or inactive
Free-ride mode is implicit always enabled (see documentation)
we only need to track when a user requests a connection.
This can be expressed by a boolean. Remove the reason enum
and rename the connect and disconnect functions.

Instead storing the user's connect request, we just track if the
session is active or not. So we move the boolean flag to the session
and get rid for the two trigger functions.

Note we don't have to deactive the session in session_disconnect()
because this will be done in cleanup_session() function which is
hash table destroy callback.