session: Assign a service to sessions
We have two conditions when we need to assign a service to a session.
Obvioulsy, the same has to be done for deassign a service from a
session but I keep this out in this commit message.

1) Whenever we create a new session, we activate it (free-ride). Since
a session can be created after service has gone online we need to go
through all services in the session_hash and match the newly created
session against all avialable sessions.

2) Whenever a service state changes, we try to assign the service
to all offline sessions.

In short, all decission which service is assigned/deassign to/from a
session happens in session.c and service.c is only in charge to handle
the auto connect.