gsupplicant: Use uint32 instead of string for wep_tx_keyidx
We use a string to pass the WEP key index value, but wpa_supplicant's
DBUS API expects a uint32. In addition, wpa_supplicant wraps strings
into double quotes, which always returns 0 when wpa_supplicant tries
to extract an integer value from it with atoi(). For this particular
case, 0 happens to be the value we want.

Fix this so that we use the correct type.