session: Remove unused NFACCT rules
We wanted to use NFACCT to do the accounting but it turnes out this
would be to expensive. First the number of rules and also the number
of reads (updates).

Besides it is currently unused, it the code is kind of fragile and it
keeps crashing ConnMan when playing with sessions.