session: Update active sessions only on Session Connect and Disconnect
When a session is created, session_activate() is called to set up the
initial session state. As the session is not yet connected, don't
update the service autoconnect with information regarding the newly
created session. When the session is disconnected or has been removed,
remember to clean up the autoconnect information.