session: On session connect check service is already available
When the session_connect() is called, the auto connect algorithm might
decide there is nothing to do because there is already a service
online. In this case the session stays offline until there is a change
via service_state_changed(). Therefore we need to go through the list
of online services and see if there is already a online service which
matches the current session configuration and select that one.