client: Don't call readline when in non-interactive mode
Don't set up readline or listen to standard in when in non-interactive
mode. Some shells seem to send garbage when redirected, and if the
device is not a tty readline goes into a CPU consuming loop. Also
print out an error if agent or vpnagent is attempted when run
non-interactively.

This is a follow-up for the issued discovered in patch
909990b531482741472824fdeea514c8d3a17eef.