service: Clear up connect logic when called from D-Bus
When connected via D-Bus, iterate only over the beginning of the list
where the services are either connected or connecting. If the interface
is not available, disconnect the other service(s) which use the
interface. Once all service(s) have been disconnected, bail out if any
of them was asynchronous and reported -EINPROGRESS. Do not force the
service to idle as it will not be possible to clean up once the
anychronous disconnect completes.