tethering: Fix issue where tether bridge may not be destroyed
Always notify __connman_tethering_set_disabled() when a technology
is no longer tethering so that the function properly counts the
number of tethering technologies. Remove the code that did
duplicate counting as it caused the bridge device not to be
removed if two or more technologies had been tethering at the
same time.