bluetooth: Create and remove PAN networks also on adapter change
Once a Bluetooth adapter changes its Powered property, PAN networks
need to be created or removed. Factor out device and PAN creation
and removal in order to call it when the adapter changes in addition
to the normal device enabling and disabling.

PAN network creation and removal on adapter change needs to be done
explicitely as the Bluetooth adapter has already changed state and
thus the bluetooth plugin will not make a D-Bus call to Bluez nor
will the callback be called where PAN network enabling and disabling
was happening.

Thanks to Krzysztof Wilk for reporting and pointing out the issue
and for the initial quick test of this fix.