gsupplicant: Add logic handling events/properties related to P2P peers
Listen to DeviceFound and DeviceLost. Get the Peer path and properties,
store them into the interface object and call the relevant peer_found/
peer_lost callbacks provided by the wifi plugin.

Due to the fact that wpa_supplicant might not (it's a bug) send the
device address, the peer identifier will be constructed using the
peer's name instead.