dhcp: Network hash table fixes
The network_table hash removal function remove_network() cannot
be used because of we might have invalidated the DHCP connection
already and then we would try to do it again when DHCP stops.

So when DHCP is stopped, we now remove the corresponding network
from the hash manually. This way we are able to unref the network
because we have a pointer to it (the pointer was lost in earlier version)
and valgrind does not report possibly lost memory for network struct
anymore.