service: Fix D-Bus return value when connecting a service
When the service is instantly connected or an error occured, remove
the pending call and report success or failure immediately to the
caller. When the connect procedure indicates -EINPROGRESS, the
caller will be informed later and no D-Bus reply is to be sent from
this function.

The service pending call needs to be set when calling
__connman_service_connect() as later stages in the code will decide
whether to ask an Agent for input based on its non-NULL value.

Reported by Aaron McCarthy <aaron.mccarthy@jolla.com>