doc: Add informations about the experimental P2P Peer services support
As soon as p2p technology is present, whether it is enabled or not, it
will be possible to register and un-register P2P Peer services. These
will be served once p2p technology gets enabled.

When scanning for Peers is possible, Peer objects will show the proposed
services of those if some are present.