wifi: Handle incoming p2p device chicken and egg problem
When connecting with P2P, a new wifi device will be created and detected
by both connman and wpa_supplicant. This last one will automatically
create an interface out of it, thus the need to hold connman creating it
itself through the DBus API.
Instead, the device is put into a pending device list, and once the
InterfaceAdded signal has been handled, we manage the other pending
devices if any as usual.