peer: Add the core implemetation of Peer service API
ConnMan will accept a p2p service registration even if the p2p
technology and the proper peer service driver are not present. It
might happen that the user will plug in a p2p enabled device at
some later point, which will trigger the low-level WiFi p2p
registration then. Only in the case when the technology and
driver are both present, the registration will wait for
confirmation from the driver.