wifi: Add the proper logic when to start/stop listening to P2P frames
This fixes an obvious bug when 2+ services are registered, wifi plugin
will blindly request the already listening interface to start listening
on a new registration. As well as it will immediatly stop the interface
forme listening at the very first unregistration.
Adding a service count inside the wifi plugin fixes it. I put this logic
in the wifi plugin since it should not be up to the core to know about
such information.