service: Don't save or load reason for failure
Loading the 'Failure' attribute and attempting to set the service state
to failure has been applied only to non-favorite services. As non-favorite
services are never autoconnected anyway, there is no change in
functionality by removing the state setting.

Also enforcing IPv4 and IPv6 state is wrong, as the respective ipconfig
states do not match after that. And when saving the 'Failure' state, the
combined service->state should have been checked in the first place.

With this change, the helper function string2error() is not used anymore
and can be removed.