wifi: scan all channels when connected
If the wifi interface is connected and a scan is requested via dbus,
only the channels from connectable services will be scanned.  This
leaves few options to fill out a UI with a fresh list of available
services when already connected.

Fix this by performing a wifi_scan_simple() when connected.