ofono: Set ipconfig off if indicated that the protocol is not used
Turn IPv6 configuration off for an IPv4 only cellular context and turn
IPv4 configuration off for an IPv6 only context. Set the configuration
method to unknown if it was previously off.

The 'Protocol' property is never sent when the modem is in active
state, thus the connected status of the modem need not be updated when
a context changed is signalled.