rtnl: Ignore temporary cloned routes
Whenever the kernel does a IPv6 route lookup, we get a netlink
"new route" notification for the destination with the RTM_F_CLONED
flag set to note that this is a temporary route.

Let's ignore these messages, or else our log may get filled with
tons of add/del route noise.