service: Let service error indicate invalid WiFi PSK passphrase
Deleting a WiFi PSK passphrase in response to an invalid key error
causes the passphrase disappear from the save file. This results in
a discovered WiFi network with identical SSID and security method
to remove the saved passphrase of the actual network.

Instead of always removing the passphrase, use service error to keep
track of the passphrase validity. Do this by unsetting service error
only when the service enters ready or disconnect states or when the
passphrase is supplied by the agent. In all other states the service
error is not modified and the error is used to detect and invalid key
upon service connection.

Reported by Pasi Sjöholm.