dnsproxy: Send pending requests to added server
When new DNS server is added, we also send all the pending
requests to this new server. Before this change we sent the
pending request to all servers even thou we had done that already.
This way we avoid sending the same DNS packet many times
to the same server.

Before this commit we also flushed the pending requests when
a DNS server was removed. This is not needed and does not make
much sense to send DNS request many times to those servers that
we are waiting response from.

Normally this commit does nothing as DNS responses come very fast
from DNS server. Because of this the request list is typically
empty so the flushing the pending request list is then a no-op.