main: Disable 6to4 by default
The 6to4 protocol as deployed on the public internet has been shown to
cause severe connectivity problems. RFC 6343 section 4.1 therefore
recommends against host implementations enabling it by default. This
patch implements that recommendation by introducing a new "Enable6to4"
configuration option, which defaults to false.

Due to the operational problems described in RFC 6343 section 3, the
IETF is about to deprecate the 6to4 protocol completely. This is done in
the document draft-ietf-v6ops-6to4-to-historic which is currently in the
RFC Editor's queue, pending final pulication as an RFC. Section 4 of
this document reinforces and strengthens RFC 6343's current
recommendation, mandating that hosts MUST disable 6to4 by default.