gsupplicant: Add interface_ap_create_fail to notify AP creation failure
Add new boolean ap_create_inprogress for GSupplicantInterface and set
this boolean to true when the wifi plugin tries to enable an AP. If
gsupplicant receives interface state information as disconnected while
ap_create_inprogress is true, this identifies there is an AP creation
failure. When that happens, gsupplicant will call the
interface_ap_create_fail callback letting the calling wifi plugin to
know about the failure.

This additional functionality is needed as the current functions may
all return success, but enabling the AP can still fail due to not being
a supported feature.