wifi: Add AP failure handling mechanism
Track the AP success or failure with a tristate variable in the wifi
plugin in order not to repeatedly try to enable AP mode on wifi
interfaces that do not support such features.

Implement callback function for the AP creation failure case and call
tech_set_tethering() in order to create an AP on the next interface
if failed.