ipconfig: Ensure ifname is not NULL when setting IPv6 state
This is a quick fix for the issue where removal of an interface may
cause IPv6 to be disabled/enabled on other interfaces.

When removing an ipdevice in free_ipdevice(), it is ensured that the
ifname (which we get from connman_inet_ifname()) is not NULL,
before invoking set_ipv6_state(). If a NULL ifname is passed into
set_ipv6_state(), it will apply the change to all interfaces instead
of the intended one.