ippool: Try to assign the same subnet while tethering
When enabling and disabling tethering, the next available /24
IPv4 subnet will be assigned and not the same one used last
time.

Fix this by starting with the last allocated block instead of
unconditionally with the next free block. This way the subnet
used for tethering will stay the same unless the subnet block
is unavailable due to other reasons. Note that even with this
change the subnet used for tethering is not guaranteed to
always be the same.

Fixes CM-666.