wifi: set interface when added
Otherwise g_supplicant_interface_set_data(interface, NULL) is not called
for the g_suplicant interface when the wifi object is removed. As a
result, the g_suplicant interface may access the wifi object when it is
already deleted.

This is rarely a problem, because the g_suplicant interface is deleted
shortly afterwards, but occasionally a D-Bus signal arrives at the wrong
time and triggers this use after free issue.