connmand-wait-online: Add command line switch for interfaces
Add command line switch to specify one ore more interfaces to wait
for before deciding if the systemd is "online" in systemd terms.

Once the global Manager state is known to be online, request all
services and iterate over the service properties to find out if
the service state is 'ready' or 'online' for a service with the
specified interface name.