dhcp: Nameserver issue on IP address renewal after lease expiry
After lease expiry and releasing its current IP, when device eventually
gets an IP address, DNS resolution stops working. This is because
nameserver list in service structure is empty. On lease expiry when
apply_dhcp_invalidate_on_network(dhcp) is called, it removes
nameservers from the service structure but it does not free namservers in
dhcp structure. Now next time when device gets an IP address and
apply_lease_available_on_network is called, nameserver is not updated as
compare_string_arrays(nameservers, dhcp->nameservers) would return true.
The fix for this would be to free dhcp->nameservers when lease is lost so
that on successful renewal next time, nameserver properly gets updated.