dhcp: Timeserver issue on IP address renewal after lease expiry
Similar to nameservers, timeservers are also checked against previous
values when added. When there is no change in timeservers when the DHCP
lease is lost and reacquired, the service will not be notified if the
timeservers have changed. With timeservers removed on lease lost,
timeservers are not properly updated once reacquired.

Fix this by removing timeservers from the dhcp struct when lease is lost.