connection: Use the sorted service list to find the default gateway
The service list is always sorted and thus the service that should have
the default gateway is first in the list. The first service in the list
is the one with the largest 'order' value and the gateway 'order'
variables all contain the same value as the respective service 'order'
variables.