gsupplicant: Mem leak in wpa_s because "RemoveNetwork" not called
ConnMan did not call netwok_remove in case AP deauthenticated client
causing wpa_s to re-allocate the ssid pointer even if the next
connection attempt is for the same SSID. This change ensures that at
the time of connection (DBUS Method call AddNetwork) if the network
is found not removed, it calls the dbus API to remove the network
and once network is removed, proceed with the connection.

By the time there is a request for connect and the network path is
not NULL it means that ConnMan has not removed the previous network
pointer. This can happen in the case AP deauthenticated client and
ConnMan does not remove the previously connected network pointer.
This causes supplicant to reallocate the memory for struct wpa_ssid
again even if it is the same SSID. This causes memory usage of
wpa_supplicnat to go high. The idea here is that if the previously
connected network is not removed at the time of next connection
attempt check if the network path is not NULL. In case it is
non-NULL first remove the network and then once removal is
successful, add the network.

Tested by running a deauth loop script at the AP end and ensure
that wpa_s does not allocate memory for struct wpa_ssid for all the
subsequent connection attempts. During the test memory usage of
wpa_s is monitored.